"amazon aws S3 done"

+++ src/resturants/resturants.controller.ts
@@ -13,6 +13,8 @@ import { CreateResturantDto } from './dto/create-resturant.dto';
 import { UpdateResturantDto } from './dto/update-resturant.dto';
 import { ResturantsService } from './resturants.service';
 import { Resturant } from './schemas/resturants.scheema';
+import { UseInterceptors, UploadedFiles } from '@nestjs/common/decorators';
+import { FilesInterceptor } from '@nestjs/platform-express';
 
 @Controller('resturants')
 export class ResturantsController {
@@ -54,12 +56,34 @@ export class ResturantsController {
     @Param('id')
     id: string,
   ): Promise<{ deleted: boolean }> {
-    const resturant = this.resturantsService.deleteById(id);
+    const resturant = await this.resturantsService.findById(id);
 
-    if (resturant) {
+    const isDeleted = await this.resturantsService.deleteImages(
+      resturant.images,
+    );
+
+    if (isDeleted) {
+      this.resturantsService.deleteById(id);
       return {
         deleted: true,
       };
+    } else {
+      return {
+        deleted: false,
+      };
     }
+    //
+  }
+
+  @Put('upload/:id')
+  @UseInterceptors(FilesInterceptor('files'))
+  async uploadFiles(
+    @Param('id') id: string,
+    @UploadedFiles() files: Array<Express.Multer.File>,
+  ) {
+    await this.resturantsService.findById(id);
+
+    const res = await this.resturantsService.uploadImages(id, files);
+    return res;
   }
 }
